CXP-599: renaming
