+++ src/Akeneo/Connectivity/Connection/back/Domain/Audit/Persistence/Query/SelectEventsApiRequestCountPerDateTime.php
@@ -9,7 +9,7 @@ namespace Akeneo\Connectivity\Connection\Domain\Audit\Persistence\Query;
  * @copyright 2020 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-interface SelectEventApiRequestsCountPerDateTime
+interface SelectEventsApiRequestCountPerDateTime
 {
     public function execute(\DateTimeImmutable $dateTime): int;
 }
