+++ src/Akeneo/Connectivity/Connection/back/tests/Integration/Fixtures/EventsApiRequestCountLoader.php
@@ -11,7 +11,7 @@ use Doctrine\DBAL\Types\Types;
  * @copyright 2020 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class EventApiRequestsCountLoader
+class EventsApiRequestCountLoader
 {
     private DbalConnection $dbalConnection;
