+++ src/Akeneo/Connectivity/Connection/back/tests/Integration/Resources/config/loaders.yml
@@ -72,6 +72,6 @@ services:
             - '@database_connection'
 
     akeneo_connectivity.connection.fixtures.event_api_request_count_loader:
-        class: Akeneo\Connectivity\Connection\back\tests\Integration\Fixtures\EventApiRequestsCountLoader
+        class: Akeneo\Connectivity\Connection\back\tests\Integration\Fixtures\EventsApiRequestCountLoader
         arguments:
             - '@database_connection'
